requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# ---------- Torch / Embeddings ----------
torch==2.1.2
sentence-transformers==2.2.2
transformers==4.41.0
huggingface-hub==0.23.0
tokenizers==0.19.1

# ---------- LangChain (LOCKED & COMPATIBLE) ----------
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-groq==0.1.4

# ---------- Vector DB ----------
faiss-cpu==1.7.4

# ---------- NLP / Utils ----------
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
nltk==3.8.1

# ---------- App ----------
streamlit==1.32.2
pypdf==4.2.0
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.4
